requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
defusedxml==0.4.1
Django==2.2
django-allauth==0.38.0
django-crispy-forms==1.7.2
idna==2.7
ipfsapi==0.4.3
mysqlclient==1.3.13
oauthlib==2.1.0
pycparser==2.19
PyMySQL==0.9.2
python3-openid==3.1.0
pytz==2018.6
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.3.0
urllib3==1.24
